Update README.md
In this assignment, you will create a music recommender system similar to the one de-scribed in chapter 5 of the textbook, so read the chapter before reading the rest of this assignment. The learning objectives of this assignment are:

Practice with imperative programming (while-and for-loops) and mutable data types (lists and dictionaries).

Practice designing a program that combines several features and needs to be im-plemented using several functions.

Important:

Do exactly as instructed in this assignment to receive full marks.

All text in this document may be considered a requirement. At no point should your program crash.

If you would like to make your program do di erent things, go for it! But include a fancy-mode switch, like the ’debug’ variable in the tictactoe game (slide 15 of Lecture 9). Use it in your program so if it’s false, the program behaves exactly as it is shown in the spec document. If fancy-mode is true, then your program can do whatever cool stu you like. When you submit, make sure fancy-mode is switched o . Write a comment in Canvas to let the CAs know that you did something super cool so we can check it out.